移除多余的 DriveMsg
