+++ TeamCode/src/main/java/org/betastudio/ftc/util/message/DriveMsg.java
@@ -1,7 +1,25 @@
 package org.betastudio.ftc.util.message;
 
-public class DriveMsg extends DriveBufMsg {
-	public DriveMsg(final double bufX, final double bufY, final double bufTurn) {
-		super(bufX, bufY, bufTurn);
+import androidx.annotation.NonNull;
+
+import java.util.Locale;
+
+public class DriveMsg implements Message {
+	public final double valX, valY, valTurn;
+
+	public DriveMsg(final double valX, final double valY, final double valTurn) {
+		this.valX = valX;
+		this.valY = valY;
+		this.valTurn = valTurn;
+	}
+
+	public DriveMsg(final double globalBuf) {
+		this(globalBuf, globalBuf, globalBuf);
+	}
+
+	@NonNull
+	@Override
+	public String toString() {
+		return String.format(Locale.SIMPLIFIED_CHINESE,"x=%.2f, y=%.2f, turn=%.2f", valX, valY, valTurn);
 	}
 }
